+++ casper/src/test/scala/coop/rchain/casper/api/ListeningNameAPITest.scala
@@ -44,7 +44,7 @@ class ListeningNameAPITest extends FlatSpec with Matchers with BlockStoreFixture
       Channel(Quote(Par().copy(exprs = Seq(Expr(GInt(2)), Expr(GInt(1)), Expr(GInt(3))))))
     val resultData = Par().copy(exprs = Seq(Expr(GInt(0))))
     val listeningNameResponse1 =
-      BlockAPI.getListeningNameDataResponse[Id](Long.MaxValue, listeningName)
+      BlockAPI.getListeningNameDataResponse[Id](Int.MaxValue, listeningName)
     val data1   = listeningNameResponse1.blockResults.map(_.postBlockData)
     val blocks1 = listeningNameResponse1.blockResults.map(_.block)
     data1 should be(List(List(resultData)))
@@ -70,7 +70,7 @@ class ListeningNameAPITest extends FlatSpec with Matchers with BlockStoreFixture
     val listeningName = Channel(Quote(Par().copy(exprs = Seq(Expr(GInt(0))))))
     val resultData    = Par().copy(exprs = Seq(Expr(GInt(0))))
     val listeningNameResponse1 =
-      BlockAPI.getListeningNameDataResponse[Id](Long.MaxValue, listeningName)
+      BlockAPI.getListeningNameDataResponse[Id](Int.MaxValue, listeningName)
     val data1   = listeningNameResponse1.blockResults.map(_.postBlockData)
     val blocks1 = listeningNameResponse1.blockResults.map(_.block)
     data1 should be(List(List(resultData)))
@@ -96,7 +96,7 @@ class ListeningNameAPITest extends FlatSpec with Matchers with BlockStoreFixture
     nodes(2).receive()
 
     val listeningNameResponse2 =
-      BlockAPI.getListeningNameDataResponse[Id](Long.MaxValue, listeningName)
+      BlockAPI.getListeningNameDataResponse[Id](Int.MaxValue, listeningName)
     val data2   = listeningNameResponse2.blockResults.map(_.postBlockData)
     val blocks2 = listeningNameResponse2.blockResults.map(_.block)
     data2 should be(
@@ -129,7 +129,7 @@ class ListeningNameAPITest extends FlatSpec with Matchers with BlockStoreFixture
     nodes(2).receive()
 
     val listeningNameResponse3 =
-      BlockAPI.getListeningNameDataResponse[Id](Long.MaxValue, listeningName)
+      BlockAPI.getListeningNameDataResponse[Id](Int.MaxValue, listeningName)
     val data3   = listeningNameResponse3.blockResults.map(_.postBlockData)
     val blocks3 = listeningNameResponse3.blockResults.map(_.block)
     data3 should be(
@@ -178,7 +178,7 @@ class ListeningNameAPITest extends FlatSpec with Matchers with BlockStoreFixture
       Some(Par().copy(exprs = Vector(Expr(GInt(0)))))
     )
     val listeningNameResponse1 =
-      BlockAPI.getListeningNameContinuationResponse[Id](Long.MaxValue, listeningNamesShuffled1)
+      BlockAPI.getListeningNameContinuationResponse[Id](Int.MaxValue, listeningNamesShuffled1)
     val continuations1 = listeningNameResponse1.blockResults.map(_.postBlockContinuations)
     val blocks1        = listeningNameResponse1.blockResults.map(_.block)
     continuations1 should be(List(List(result)))
@@ -191,7 +191,7 @@ class ListeningNameAPITest extends FlatSpec with Matchers with BlockStoreFixture
         Channel(Quote(Par().copy(exprs = Seq(Expr(GInt(1)), Expr(GInt(2))))))
       )
     val listeningNameResponse2 =
-      BlockAPI.getListeningNameContinuationResponse[Id](Long.MaxValue, listeningNamesShuffled2)
+      BlockAPI.getListeningNameContinuationResponse[Id](Int.MaxValue, listeningNamesShuffled2)
     val continuations2 = listeningNameResponse2.blockResults.map(_.postBlockContinuations)
     val blocks2        = listeningNameResponse2.blockResults.map(_.block)
     continuations2 should be(List(List(result)))
